# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PropTech Platform API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # COMPANIES
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # USERS
...
  /counterparties:
    get:
      operationId: listCounterparties
      tags: [counterparties]
      summary: List counterparties
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of counterparties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Counterparty"
    post:
      operationId: createCounterparty
      tags: [counterparties]
      summary: Create counterparty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCounterpartyBody"
      responses:
        "201":
          description: Counterparty created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counterparty"
  /counterparties/{id}:
    get:
      operationId: getCounterparty
      tags: [counterparties]
      summary: Get counterparty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Counterparty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counterparty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCounterparty
      tags: [counterparties]
      summary: Update counterparty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCounterpartyBody"
      responses:
        "200":
          description: Updated counterparty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counterparty"
    delete:
      operationId: deleteCounterparty
      tags: [counterparties]
      summary: Delete counterparty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PROPERTIES
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: counterpartyId
          in: query
          schema:
            type: integer
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DOCUMENTS
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # IMPORT
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        legalName:
          type: ["string", "null"]
        bin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - isActive
        - createdAt
        - updatedAt
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        legalName:
          type: ["string", "null"]
        bin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required:
        - name
    UpdateCompanyBody:
      type: object
      properties:
        name:
          type: string
        legalName:
          type: ["string", "null"]
        bin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        isActive:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, company_admin, sales_manager, finance, rental_manager, staff]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
        - updatedAt
...
    Counterparty:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [individual, legal_entity, owner, tenant]
        fullName:
          type: string
        iin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        additionalContact:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        sourceType:
          type: ["string", "null"]
        syncStatus:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - type
        - fullName
        - createdAt
        - updatedAt
    CreateCounterpartyBody:
      type: object
      properties:
        type:
          type: string
          enum: [individual, legal_entity, owner, tenant]
        fullName:
          type: string
        iin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        additionalContact:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
      required:
        - type
        - fullName
    UpdateCounterpartyBody:
      type: object
      properties:
        type:
          type: string
          enum: [individual, legal_entity, owner, tenant]
        fullName:
          type: string
        iin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        additionalContact:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        contractNumber:
          type: string
        contractDate:
          type: ["string", "null"]
        type:
          type: string
          enum: [sale, rental, service, other]
        counterpartyId:
          type: ["integer", "null"]
        counterpartyName:
          type: ["string", "null"]
        propertyId:
          type: ["integer", "null"]
        propertyUnitNumber:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        accrualDate:
          type: ["string", "null"]
        deposit:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, active, expired, terminated, archived]
        comment:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        sourceType:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - contractNumber
        - type
        - status
        - createdAt
        - updatedAt
    CreateContractBody:
      type: object
      properties:
        contractNumber:
          type: string
        contractDate:
          type: ["string", "null"]
        type:
          type: string
          enum: [sale, rental, service, other]
        counterpartyId:
          type: ["integer", "null"]
        propertyId:
          type: ["integer", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        accrualDate:
          type: ["string", "null"]
        deposit:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, active, expired, terminated, archived]
        comment:
          type: ["string", "null"]
      required:
        - contractNumber
        - type
        - status
    UpdateContractBody:
      type: object
      properties:
        contractNumber:
          type: string
        contractDate:
          type: ["string", "null"]
        type:
          type: string
          enum: [sale, rental, service, other]
        counterpartyId:
          type: ["integer", "null"]
        propertyId:
          type: ["integer", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        accrualDate:
          type: ["string", "null"]
        deposit:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, active, expired, terminated, archived]
        comment:
          type: ["string", "null"]
    Document:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        name:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        uploadedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - entityType
        - entityId
        - name
        - fileUrl
        - createdAt
    CreateDocumentBody:
      type: object
      properties:
        entityType:
          type: string
        entityId:
          type: integer
        name:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
      required:
        - entityType
        - entityId
        - name
        - fileUrl
...